Refactors attackby so that base proc call atom/attackby is always called
- Refactors ignition to be handled by two elements (igniter and
ignitable)

- Refactors writing to be handled by two element (writer and writable)

- Where applicable, moved functionality from child types to their
parents when calling of parent attackby was conditional (such as storing
lockable backpacks or unpacking crates)

- Fixed runtime caused by cryoing monkeys, farwas, and other
human-subtyped mobs without jobs

- Fixed supplyspawner landmark

+++ code/game/objects/items/frames/frame.dm
@@ -14,8 +14,12 @@
 	flags_atom = FPRINT|CONDUCT
 
 /obj/item/frame/apc/attackby(obj/item/W as obj, mob/user as mob)
-	..()
+	. = ..()
+	if (. & ATTACK_HINT_BREAK_ATTACK)
+		return
+
 	if (HAS_TRAIT(W, TRAIT_TOOL_WRENCH))
+		. |= ATTACK_HINT_NO_TELEGRAPH
 		new /obj/item/stack/sheet/metal( get_turf(src.loc), 2 )
 		qdel(src)
